Updates the FSM
I had an error where if you cheated it would bring you to the score screen and display a message that you cheated.

This is not in accordance with the lab directive. According to the instructions you should be navigated back to the starting screen after cheating. I have made changes in the code such that the cheating condition redirects to the reset state.