Adds preliminary change time functionality
In reality, this push contains a lot of broken code. As of now, there is weird behaviour with the changing time function. I can iterate through the individual states however I can't seem to change any of the dates. I think I'll go eat some dinner and come back to this with a fresh mind.