Adds ADC functionality
Through a painful process, I have figured out that it is impossible to used the ADC in either DMA or interrupt mode. DMA is impossible because the register for DMA is apparently used by the LCD screen and you can't use interrupts because the interrupt will interrupt your code too frequently stopping it from checking if the button was held for 0.5 seconds.

Right now the code is really messy but it works. I'll make another commit once I've cleaned up the source code.