Configures motor GPIO
This commit configures the GPIO pins in a push pull configuration to drive the two solenoid pairs of A1A2, B1B2.

Through testing I have confirmed that the motor runs, it switches direction and between the two stepping modes. The answer for both the demo questions is "The motor will switch directions from it's intended spinning direction".

All that is left to do is for Fil to implement the change speed function.